move remaining adapters exception imports to common/adapters
