+++ core/dbt/clients/jinja_static.py
@@ -1,6 +1,7 @@
 import jinja2
 from dbt.clients.jinja import get_environment
-from dbt.exceptions import MacroNamespaceNotStringError, MacroNameNotStringError
+from dbt.exceptions import MacroNamespaceNotStringError
+from dbt.common.exceptions.macros import MacroNameNotStringError
 
 
 def statically_extract_macro_calls(string, ctx, db_wrapper=None):
