+++ core/dbt/clients/yaml_helper.py
@@ -1,3 +1,4 @@
+import dbt.common.exceptions.base
 import dbt.exceptions
 from typing import Any, Dict, Optional
 import yaml
@@ -60,4 +61,4 @@ def load_yaml_text(contents, path=None):
         else:
             error = str(e)
 
-        raise dbt.exceptions.DbtValidationError(error)
+        raise dbt.common.exceptions.base.DbtValidationError(error)
