+++ core/dbt/common/exceptions/__init__.py
@@ -1,2 +1,3 @@
 from dbt.common.exceptions.base import *  # noqa
 from dbt.common.exceptions.events import *  # noqa
+from dbt.common.exceptions.macros import *  # noqa
