+++ core/dbt/config/profile.py
@@ -13,10 +13,10 @@ from dbt.exceptions import (
     CompilationError,
     DbtProfileError,
     DbtProjectError,
-    DbtValidationError,
     DbtRuntimeError,
     ProfileConfigError,
 )
+from ..common.exceptions import DbtValidationError
 from dbt.common.events.types import MissingProfileTarget
 from dbt.common.events.functions import fire_event
 from dbt.utils import coerce_dict_str
