+++ core/dbt/config/utils.py
@@ -4,7 +4,8 @@ from typing import Any, Dict
 from dbt.clients import yaml_helper
 from dbt.common.events.functions import fire_event
 from dbt.common.events.types import InvalidOptionYAML
-from dbt.exceptions import DbtValidationError, OptionNotYamlDictError
+from dbt.exceptions import OptionNotYamlDictError
+from dbt.common.exceptions import DbtValidationError
 
 
 def parse_cli_vars(var_string: str) -> Dict[str, Any]:
