+++ core/dbt/context/base.py
@@ -16,11 +16,11 @@ from dbt.contracts.graph.nodes import Resource
 from dbt.exceptions import (
     SecretEnvVarLocationError,
     EnvVarMissingError,
-    MacroReturn,
     RequiredVarNotFoundError,
     SetStrictWrongTypeError,
     ZipStrictWrongTypeError,
 )
+from dbt.common.exceptions.macros import MacroReturn
 from dbt.common.events.functions import fire_event, get_invocation_id
 from dbt.common.events.types import JinjaLogInfo, JinjaLogDebug
 from dbt.common.events.contextvars import get_node_info
