+++ core/dbt/context/exceptions_jinja.py
@@ -4,31 +4,30 @@ from typing import NoReturn
 from dbt.common.events.functions import warn_or_error
 from dbt.common.events.types import JinjaLogWarning
 
-from dbt.common.exceptions import DbtRuntimeError, NotImplementedError
+from dbt.common.exceptions import DbtRuntimeError, NotImplementedError, DbtDatabaseError
 from dbt.adapters.exceptions import (
     MissingConfigError,
     ColumnTypeMissingError,
     MissingMaterializationError,
+    RelationWrongTypeError,
 )
 from dbt.exceptions import (
     MissingRelationError,
     AmbiguousAliasError,
     AmbiguousCatalogMatchError,
-    CacheInconsistencyError,
     DataclassNotDictError,
     CompilationError,
-    DbtDatabaseError,
     DependencyNotFoundError,
     DependencyError,
     DuplicatePatchPathError,
     DuplicateResourceNameError,
     PropertyYMLError,
-    RelationWrongTypeError,
     ContractError,
     FailFastError,
     scrub_secrets,
     env_secrets,
 )
+from dbt.common.exceptions.cache import CacheInconsistencyError
 
 
 def warn(msg, node=None):
