+++ core/dbt/context/providers.py
@@ -43,7 +43,7 @@ from dbt.contracts.graph.nodes import (
 from dbt.contracts.graph.metrics import MetricReference, ResolvedMetricReference
 from dbt.contracts.graph.unparsed import NodeVersion
 from dbt.common.events.functions import get_metadata_vars
-from dbt.common.exceptions import DbtInternalError, DbtRuntimeError
+from dbt.common.exceptions import DbtInternalError, DbtRuntimeError, DbtValidationError
 from dbt.adapters.exceptions import MissingConfigError
 from dbt.exceptions import (
     CompilationError,
@@ -57,7 +57,6 @@ from dbt.exceptions import (
     LoadAgateTableValueError,
     MacroDispatchArgError,
     MacroResultAlreadyLoadedError,
-    MacrosSourcesUnWriteableError,
     MetricArgsError,
     OperationsCannotRefEphemeralNodesError,
     PackageNotInDepsError,
@@ -65,9 +64,9 @@ from dbt.exceptions import (
     RefBadContextError,
     RefArgsError,
     TargetNotFoundError,
-    DbtValidationError,
     DbtReferenceError,
 )
+from dbt.common.exceptions.macros import MacrosSourcesUnWriteableError
 from dbt.config import IsFQNResource
 from dbt.node_types import NodeType, ModelLanguage
