+++ tests/unit/test_postgres_adapter.py
@@ -16,7 +16,8 @@ from dbt.adapters.postgres import Plugin as PostgresPlugin
 from dbt.contracts.files import FileHash
 from dbt.contracts.graph.manifest import ManifestStateCheck
 from dbt.common.clients import agate_helper
-from dbt.exceptions import DbtValidationError, DbtConfigError
+from dbt.exceptions import DbtConfigError
+from dbt.common.exceptions import DbtValidationError
 from psycopg2 import extensions as psycopg2_extensions
 from psycopg2 import DatabaseError
